refactor: rename ad_content -> ad_description and trim superficial tests
- Rename module and test file for clearer domain boundary
- Remove 21 superficial/duplicate tests per oracle review
- Keep high-signal tests: affix precedence, @ sanitization,
  condition normalization, xpath_literal, marker extraction
